Move AllowedSecondFactorMap out of domain
